# swagger.yaml
openapi: 3.0.0
info:
  title: Sansal AI Interview API
  description: API documentation for the Sansal AI Interview Platform.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local Development Server
  - url: https://sansal-backend-xyzsandeepsansal.onrender.com/api
    description: Production Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
        type:
          type: string
          enum: [booking, meeting_started, system]
        title:
          type: string
        message:
          type: string
        data:
          type: object
        isRead:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        isCleared:
          type: boolean
        clearedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    Note:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        tags:
          type: array
          items:
            type: string
        fileUrl:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        uploader:
          type: object
          properties:
            id:
              type: string
              nullable: true
            name:
              type: string
            role:
              type: string
              enum: [student, expert, admin]
        createdAt:
          type: string
          format: date-time
paths:
...
  /users/interviews:
    get:
      summary: Get interview history
      tags: [Users]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
        - in: query
          name: sortBy
          schema: { type: string, default: "createdAt:desc" }
      responses:
        '200':
          description: History retrieved
  /users/interviews/{id}:
    get:
      summary: Get specific interview details
      tags: [Users]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Interview details retrieved
...
  /notifications/{id}:
    delete:
      summary: Clear a notification
      tags: [Notifications, Expert]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Notification cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  notification:
                    $ref: '#/components/schemas/Notification'
...
  /notes:
    get:
      summary: List public notes
      tags: [Notes]
      security: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search query (title/description/subject/tags)
        - in: query
          name: subject
          schema:
            type: string
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: uploader
          schema:
            type: string
            enum: [student, expert, admin]
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest, title]
      responses:
        '200':
          description: Notes retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Note'
    post:
      summary: Upload a note
      tags: [Notes]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, file]
              properties:
                title:
                  type: string
                description:
                  type: string
                subject:
                  type: string
                tags:
                  type: string
                  description: Comma separated tags
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Note uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    $ref: '#/components/schemas/Note'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Server error

  # -------------------------
  # Interview Routes (/api/interviews)
  # -------------------------
  /interviews:
    post:
      summary: Create interview (Interview Controller)
      tags: [Interviews]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  type: object
                  default:
                    role: "devops"
                    experience: "senior"
                    difficulty: "easy"
                    duration: 45
                questions:
                  type: array
                  default:
                    - id: 1
                      question: "What is CI/CD and why is it important?"
                      ideal_answer: "CI/CD stands for Continuous Integration and Continuous Deployment. It's important because it automates the software delivery process, reduces errors, and enables faster release cycles."
                      category: "DevOps Concepts"
                      difficulty: "Easy"
                      role: "DevOps Engineer"
                      experience: "fresher"
                      timeLimit: 120
                answers:
                  type: array
                  default:
                    - question:
                        id: 1
                        question: "What is CI/CD and why is it important?"
                        ideal_answer: "CI/CD stands for Continuous Integration and Continuous Deployment. It's important because it automates the software delivery process, reduces errors, and enables faster release cycles."
                        category: "DevOps Concepts"
                        difficulty: "Easy"
                        role: "DevOps Engineer"
                        experience: "fresher"
                        timeLimit: 120
                      answer: "dasdas"
                      score: 0
                      timeSpent: 5
                      skipped: false
                      timestamp: "2026-01-19T19:18:20.756Z"
                totalScore: { type: number, default: 0 }
                timeSpent: { type: integer, default: 7 }
                completed: { type: boolean, default: true }
                timestamp: { type: string, default: "2026-01-19T19:18:21.662Z" }
      responses:
        '201':
          description: Interview created
    get:
      summary: Get all interviews for user
      tags: [Interviews]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of interviews
...